requirements.txt
Flask==2.0.1
gunicorn==20.1.0
google-cloud-storage==1.42.0
uvicorn==0.15.0
fastapi==0.68.1
google-cloud-aiplatform==1.4.0